Date 관련 컴포넌트 개편 (#334)
- SessionTime을 String Type -> LocalDateTime으로 변경
- DateUtil제거
- UI에 존재하는 시간 관련 계산로직 제거

+++ data/src/main/java/com/yapp/data/datasource/SessionRemoteDataSourceImpl.kt
@@ -26,7 +26,7 @@ class SessionRemoteDataSourceImpl @Inject constructor(
         }
     }
 
-    override suspend fun getSession(id: Long): SessionEntity? {
+    override suspend fun getSession(id: Int): SessionEntity? {
         return suspendCancellableCoroutine { cancellableContinuation ->
             fireStore.sessionRef()
                 .document(id.toString())
